content/media/video/test/test_async_load_source_append.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478299
-->
<head>
  <title>Test for Bug 478299</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478299">Mozilla Bug 478299</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 478299 **/

// This tests that we only uses sources added before a load is triggered in an async load.
// See: https://bugzilla.mozilla.org/show_bug.cgi?id=465458#c4

var gLoadStartsCount = 0;
var gLoadedCount = 0;
var gAbortCount = 0;
var gEmptiedCount = 0;
var gErrorCount = 0;

function errorHandler(e) {
  gErrorCount++;
  is(gLoadStartsCount, 1, "Should have received 1 loadstart.");
  is(gLoadedCount, 0, "Should not have loaded");
  is(gAbortCount, 0, "Shouldn't have aborted any loads");
  is(gErrorCount, 1, "Shouldn't have received errors");
  is(gEmptiedCount, 0, "Shouldn't have reEmptiedCount should be 1");
  SimpleTest.finish();  
}

function abortHandler() {
  gAbortCount++;
}

function startHandler() {
  gLoadStartsCount++;
}

function loadedHandler() {
  gLoadedCount++;
}

function emptiedHandler() {
  gEmptiedCount++;
}

function volumeChangeHandler() {
  gVolumeChanged++;
}

var v = document.createElement('video');
v.addEventListener('error', errorHandler, false);
v.addEventListener('abort', abortHandler, false);
v.addEventListener('loadeddata', loadedHandler, false);
v.addEventListener('loadstart', startHandler, false);
v.addEventListener('emptied', emptiedHandler, false);

// Create a source to a 404. Expect error on load.
var s1 = document.createElement('source');
s1.type = "application/ogg";
s1.src = "404.ogv";
v.appendChild(s1);
v.load();

// Create a source to a valid resource.
var s2 = document.createElement('source');
s2.type = "application/ogg";
s2.src = "320x240.ogv";

// We expect that this source will not be considered during load.
v.appendChild(s2);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
